services/horizon/internal/db2/history: Fix fee bump migrations (#2429)
* Change type of fee_charged column to biginteger

The XDR definiton of fee_charged assumes a 64 bit representation. However, the database representation in Horizon has always been
a 32 bit integer because it was only possible to set a 32 bit fee on a transaction. In protocol 13, it will be possible to
set a 64 bit fee in a fee bump transaction. We must change the column type of fee_charged to biginteger before protocol 13 is released.

Unfortunately, changing the type of the column will trigger a rewrite of the history_transactions table.
Note that multiple columns are changed using a single ALTER TABLE statement. According to https://www.postgresql.org/docs/9.6/sql-altertable.html
when multiple columns are changed using a single statement only one pass through the table is required to perform the operation:
"The main reason for providing the option to specify multiple changes in a single ALTER TABLE is that multiple table scans or rewrites can thereby be combined into a single pass over the table."

+++ protocols/horizon/main.go
@@ -432,17 +432,20 @@ type Transaction struct {
 		Precedes   hal.Link `json:"precedes"`
 		Succeeds   hal.Link `json:"succeeds"`
 	} `json:"_links"`
-	ID                 string              `json:"id"`
-	PT                 string              `json:"paging_token"`
-	Successful         bool                `json:"successful"`
-	Hash               string              `json:"hash"`
-	Ledger             int32               `json:"ledger"`
-	LedgerCloseTime    time.Time           `json:"created_at"`
-	Account            string              `json:"source_account"`
-	AccountSequence    string              `json:"source_account_sequence"`
-	FeeAccount         string              `json:"fee_account"`
-	FeeCharged         int32               `json:"fee_charged"`
-	MaxFee             int32               `json:"max_fee"`
+	ID              string    `json:"id"`
+	PT              string    `json:"paging_token"`
+	Successful      bool      `json:"successful"`
+	Hash            string    `json:"hash"`
+	Ledger          int32     `json:"ledger"`
+	LedgerCloseTime time.Time `json:"created_at"`
+	Account         string    `json:"source_account"`
+	AccountSequence string    `json:"source_account_sequence"`
+	FeeAccount      string    `json:"fee_account"`
+	// Action needed in release: horizonclient-v3.0.0
+	// Change FeeCharged and MaxFee types to string
+	// because JSON doesn't support 64 bit integers
+	FeeCharged         int64               `json:"fee_charged"`
+	MaxFee             int64               `json:"max_fee"`
 	OperationCount     int32               `json:"operation_count"`
 	EnvelopeXdr        string              `json:"envelope_xdr"`
 	ResultXdr          string              `json:"result_xdr"`
